backend/requirements.txt
flask
flask-cors
groq
scikit-learn
numpy
python-dotenv
